+++ src/eliottree/test/tasks.py
@@ -28,14 +28,14 @@ nested_action_task = {
     u"task_level": [1, 1]}
 
 action_task_end = {
-    u"timestamp": 1425356800,
+    u"timestamp": 1425356802,
     u"action_status": u"succeeded",
     u"task_uuid": u"f3a32bb3-ea6b-457c-aa99-08a3d0491ab4",
     u"action_type": u"app:action",
     u"task_level": [2]}
 
 action_task_end_failed = {
-    u"timestamp": 1425356800,
+    u"timestamp": 1425356804,
     u"action_status": u"failed",
     u"task_uuid": u"f3a32bb3-ea6b-457c-aa99-08a3d0491ab4",
     u"action_type": u"app:action",
@@ -72,7 +72,6 @@ janky_action_task = {
     u"action_type": u"A\x1b(0",
     u"task_level": [1],
     u"mes\nsage": u"hello\x1b(0world",
-    u"\x1b(0": "nope",
     u"\x1b(0": {u"\x1b(0": "nope"}}
 
 janky_message_task = {