+++ src/eliottree/test/test_render.py
@@ -4,13 +4,15 @@ from termcolor import colored
 from testtools import ExpectedException, TestCase
 from testtools.matchers import AfterPreprocessing as After
 from testtools.matchers import (
-    Contains, EndsWith, Equals, HasLength, IsDeprecated, StartsWith)
+    Contains, Equals, HasLength, IsDeprecated, MatchesAll, MatchesListwise,
+    StartsWith)
 
 from eliottree import (
     Tree, render_task_nodes, render_tasks, tasks_from_iterable)
 from eliottree._render import (
-    COLORS, RIGHT_DOUBLE_ARROW, _default_value_formatter, _no_color,
+    COLORS, HOURGLASS, RIGHT_DOUBLE_ARROW, _default_value_formatter, _no_color,
     format_node, get_children, message_fields, message_name)
+from eliottree._util import eliot_ns
 from eliottree.render import get_name_factory
 from eliottree.test.matchers import ExactlyEquals
 from eliottree.test.tasks import (
@@ -72,7 +74,7 @@ class DefaultValueFormatterTests(TestCase):
 
     def test_timestamp_field(self):
         """
-        Format ``timestamp`` fields as human-readable if the feature was
+        Format Eliot ``timestamp`` fields as human-readable if the feature was
         requested.
         """
         format_value = _default_value_formatter(
@@ -80,11 +82,25 @@ class DefaultValueFormatterTests(TestCase):
         # datetime(2015, 6, 6, 22, 57, 12)
         now = 1433631432
         self.assertThat(
-            format_value(now, u'timestamp'),
+            format_value(now, eliot_ns(u'timestamp')),
             ExactlyEquals(u'2015-06-06 22:57:12'))
         self.assertThat(
-            format_value(str(now), u'timestamp'),
+            format_value(str(now), eliot_ns(u'timestamp')),
             ExactlyEquals(u'2015-06-06 22:57:12'))
+
+    def test_not_eliot_timestamp_field(self):
+        """
+        Do not format user fields named ``timestamp``.
+        """
+        format_value = _default_value_formatter(
+            human_readable=True, field_limit=0)
+        now = 1433631432
+        self.assertThat(
+            format_value(now, u'timestamp'),
+            ExactlyEquals(text_type(now)))
+        self.assertThat(
+            format_value(text_type(now), u'timestamp'),
+            ExactlyEquals(text_type(now)))
 
     def test_timestamp_field_not_human(self):
         """
@@ -131,7 +147,7 @@ class RenderTaskNodesTests(TestCase):
                 u'\u2514\u2500\u2500 app:action@1/started\n'
                 u'    \u251c\u2500\u2500 timestamp: 1425356800\n'
                 u'    \u2514\u2500\u2500 app:action@2/succeeded\n'
-                u'        \u2514\u2500\u2500 timestamp: 1425356800\n\n'
+                u'        \u2514\u2500\u2500 timestamp: 1425356802\n\n'
                 .encode('utf-8')))
 
     def test_tasks_human_readable(self):
@@ -154,7 +170,7 @@ class RenderTaskNodesTests(TestCase):
                 u'\u2514\u2500\u2500 app:action@1/started\n'
                 u'    \u251c\u2500\u2500 timestamp: 2015-03-03 04:26:40\n'
                 u'    \u2514\u2500\u2500 app:action@2/succeeded\n'
-                u'        \u2514\u2500\u2500 timestamp: 2015-03-03 04:26:40\n'
+                u'        \u2514\u2500\u2500 timestamp: 2015-03-03 04:26:42\n'
                 u'\n'
                 .encode('utf-8')))
 
@@ -377,7 +393,7 @@ class RenderTaskNodesTests(TestCase):
                     u'    \u2514\u2500\u2500 {}'.format(
                         C.success(u'app:action@2/succeeded')),
                     u'        \u2514\u2500\u2500 {}: {}'.format(
-                        C.prop('timestamp'), u'1425356800'),
+                        C.prop('timestamp'), u'1425356802'),
                     u'\n',
                 ]).encode('utf-8')))
 
@@ -475,6 +491,10 @@ colors = COLORS(colored)
 no_colors = COLORS(_no_color)
 
 
+def no_formatting(value, field_name=None):
+    return text_type(value)
+
+
 class MessageNameTests(TestCase):
     """
     Tests for `eliottree.render.message_name`.
@@ -485,7 +505,7 @@ class MessageNameTests(TestCase):
         """
         message = next(tasks_from_iterable([action_task])).root().start_message
         self.assertThat(
-            message_name(colors, message),
+            message_name(colors, no_formatting, message),
             StartsWith(colors.parent(message.contents.action_type)))
 
     def test_action_task_level(self):
@@ -494,7 +514,7 @@ class MessageNameTests(TestCase):
         """
         message = next(tasks_from_iterable([action_task])).root().start_message
         self.assertThat(
-            message_name(colors, message),
+            message_name(colors, no_formatting, message),
             Contains(message.task_level.to_string()))
 
     def test_action_status(self):
@@ -503,8 +523,8 @@ class MessageNameTests(TestCase):
         """
         message = next(tasks_from_iterable([action_task])).root().start_message
         self.assertThat(
-            message_name(colors, message),
-            EndsWith(u'started'))
+            message_name(colors, no_formatting, message),
+            Contains(u'started'))
 
     def test_action_status_success(self):
         """
@@ -514,8 +534,8 @@ class MessageNameTests(TestCase):
             action_task, action_task_end,
         ])).root().end_message
         self.assertThat(
-            message_name(colors, message),
-            EndsWith(colors.success(u'succeeded')))
+            message_name(colors, no_formatting, message),
+            Contains(colors.success(u'succeeded')))
 
     def test_action_status_failed(self):
         """
@@ -525,8 +545,8 @@ class MessageNameTests(TestCase):
             action_task, action_task_end_failed,
         ])).root().end_message
         self.assertThat(
-            message_name(colors, message),
-            EndsWith(colors.failure(u'failed')))
+            message_name(colors, no_formatting, message),
+            Contains(colors.failure(u'failed')))
 
     def test_message_type(self):
         """
@@ -534,7 +554,7 @@ class MessageNameTests(TestCase):
         """
         message = WrittenMessage.from_dict(message_task)
         self.assertThat(
-            message_name(colors, message),
+            message_name(colors, no_formatting, message),
             StartsWith(colors.parent(message.contents.message_type)))
 
     def test_message_task_level(self):
@@ -543,7 +563,7 @@ class MessageNameTests(TestCase):
         """
         message = WrittenMessage.from_dict(message_task)
         self.assertThat(
-            message_name(colors, message),
+            message_name(colors, no_formatting, message),
             Contains(message.task_level.to_string()))
 
     def test_unknown(self):
@@ -552,11 +572,11 @@ class MessageNameTests(TestCase):
         ``<unnamed>``.
         """
         self.assertThat(
-            message_name(colors, None),
+            message_name(colors, no_formatting, None),
             ExactlyEquals(u'<unnamed>'))
-        message = WrittenMessage.from_dict({})
+        message = WrittenMessage.from_dict({u'timestamp': 0})
         self.assertThat(
-            message_name(colors, message),
+            message_name(colors, no_formatting, message),
             ExactlyEquals(u'<unnamed>'))
 
 
@@ -566,7 +586,7 @@ class FormatNodeTests(TestCase):
     """
     def format_node(self, node, format_value=None, colors=no_colors):
         if format_value is None:
-            def format_value(value, _):
+            def format_value(value, field_name=None):
                 return value
         return format_node(format_value, colors, node)
 
@@ -583,14 +603,17 @@ class FormatNodeTests(TestCase):
         """
         `WrittenAction`'s start message is rendered.
         """
-        node = next(tasks_from_iterable([action_task])).root()
+        tasks = tasks_from_iterable([action_task, action_task_end])
+        node = next(tasks).root()
         self.assertThat(
             self.format_node(node),
-            ExactlyEquals(u'{}{} {} {}'.format(
+            ExactlyEquals(u'{}{} {} {} {} \u29d6 {}'.format(
                 node.start_message.contents.action_type,
                 node.start_message.task_level.to_string(),
                 RIGHT_DOUBLE_ARROW,
-                node.start_message.contents.action_status)))
+                node.start_message.contents.action_status,
+                node.start_message.timestamp,
+                node.end_message.timestamp - node.start_message.timestamp)))
 
     def test_tuple_list(self):
         """
@@ -659,7 +682,7 @@ class MessageFieldsTests(TestCase):
 
     def test_fields(self):
         """
-        Include all the message fields and the timestamp.
+        Include all the message fields but not the timestamp.
         """
         message = WrittenMessage.from_dict({u'a': 1})
         self.assertThat(
@@ -668,8 +691,8 @@ class MessageFieldsTests(TestCase):
         message = WrittenMessage.from_dict({u'a': 1, u'timestamp': 12345678})
         self.assertThat(
             message_fields(message, set()),
-            Equals([(u'a', 1),
-                    (u'timestamp', 12345678)]))
+            Equals([
+                (u'a', 1)]))
 
     def test_ignored_fields(self):
         """
@@ -710,8 +733,7 @@ class GetChildrenTests(TestCase):
             list(get_children({u'c'}, node)),
             Equals([
                 (u'action_status', start_message.contents.action_status),
-                (u'action_type', start_message.contents.action_type),
-                (u'timestamp', start_message.timestamp)]))
+                (u'action_type', start_message.contents.action_type)]))
 
     def test_written_action_start(self):
         """
@@ -722,11 +744,10 @@ class GetChildrenTests(TestCase):
             action_task, nested_action_task, action_task_end])).root()
         start_message = node.start_message
         self.assertThat(
-            list(get_children({u'foo'}, node))[:3],
+            list(get_children({u'foo'}, node))[:2],
             Equals([
                 (u'action_status', start_message.contents.action_status),
-                (u'action_type', start_message.contents.action_type),
-                (u'timestamp', start_message.timestamp)]))
+                (u'action_type', start_message.contents.action_type)]))
 
     def test_written_action_children(self):
         """
@@ -735,7 +756,7 @@ class GetChildrenTests(TestCase):
         node = next(tasks_from_iterable([
             action_task, nested_action_task, action_task_end])).root()
         self.assertThat(
-            list(get_children({u'foo'}, node))[3],
+            list(get_children({u'foo'}, node))[2],
             Equals(node.children[0]))
 
     def test_written_action_no_children(self):
@@ -746,7 +767,7 @@ class GetChildrenTests(TestCase):
         node = next(tasks_from_iterable([action_task])).root()
         self.assertThat(
             list(get_children({u'foo'}, node)),
-            HasLength(3))
+            HasLength(2))
 
     def test_written_action_no_end(self):
         """
@@ -765,7 +786,7 @@ class GetChildrenTests(TestCase):
         node = next(tasks_from_iterable([
             action_task, nested_action_task, action_task_end])).root()
         self.assertThat(
-            list(get_children({u'foo'}, node))[4:],
+            list(get_children({u'foo'}, node))[3:],
             Equals([node.end_message]))
 
     def test_written_message(self):
@@ -822,9 +843,46 @@ class RenderTasksTests(TestCase):
     """
     def render_tasks(self, iterable, **kw):
         fd = StringIO()
+        err = StringIO(u'')
         tasks = tasks_from_iterable(iterable)
-        render_tasks(write=fd.write, tasks=tasks, **kw)
+        render_tasks(write=fd.write, write_err=err.write, tasks=tasks, **kw)
+        if err.tell():
+            return fd.getvalue(), err.getvalue()
         return fd.getvalue()
+
+    def test_format_node_failures(self):
+        """
+        Catch exceptions when formatting nodes and display a message without
+        interrupting the processing of tasks. List all caught exceptions to
+        stderr.
+        """
+        def bad_format_node(*a, **kw):
+            raise ValueError('Nope')
+        self.assertThat(
+            self.render_tasks([message_task],
+                              format_node=bad_format_node),
+            MatchesListwise([
+                Contains(u'<node formatting exception>'),
+                MatchesAll(
+                    Contains(u'Traceback (most recent call last):'),
+                    Contains(u'ValueError: Nope'))]))
+
+    def test_format_value_failures(self):
+        """
+        Catch exceptions when formatting node values and display a message
+        without interrupting the processing of tasks. List all caught
+        exceptions to stderr.
+        """
+        def bad_format_value(*a, **kw):
+            raise ValueError('Nope')
+        self.assertThat(
+            self.render_tasks([message_task],
+                              format_value=bad_format_value),
+            MatchesListwise([
+                Contains(u'message: <value formatting exception>'),
+                MatchesAll(
+                    Contains(u'Traceback (most recent call last):'),
+                    Contains(u'ValueError: Nope'))]))
 
     def test_tasks(self):
         """
@@ -835,10 +893,10 @@ class RenderTasksTests(TestCase):
             self.render_tasks([action_task, action_task_end]),
             ExactlyEquals(
                 u'f3a32bb3-ea6b-457c-aa99-08a3d0491ab4\n'
-                u'\u2514\u2500\u2500 app:action/1 \u21d2 started\n'
-                u'    \u251c\u2500\u2500 timestamp: 1425356800\n'
-                u'    \u2514\u2500\u2500 app:action/2 \u21d2 succeeded\n'
-                u'        \u2514\u2500\u2500 timestamp: 1425356800\n\n'))
+                u'\u2514\u2500\u2500 app:action/1 \u21d2 started '
+                u'1425356800 \u29d6 2\n'
+                u'    \u2514\u2500\u2500 app:action/2 \u21d2 succeeded '
+                u'1425356802\n\n'))
 
     def test_tasks_human_readable(self):
         """
@@ -850,10 +908,10 @@ class RenderTasksTests(TestCase):
                               human_readable=True),
             ExactlyEquals(
                 u'f3a32bb3-ea6b-457c-aa99-08a3d0491ab4\n'
-                u'\u2514\u2500\u2500 app:action/1 \u21d2 started\n'
-                u'    \u251c\u2500\u2500 timestamp: 2015-03-03 04:26:40\n'
-                u'    \u2514\u2500\u2500 app:action/2 \u21d2 succeeded\n'
-                u'        \u2514\u2500\u2500 timestamp: 2015-03-03 04:26:40\n'
+                u'\u2514\u2500\u2500 app:action/1 \u21d2 started '
+                u'2015-03-03 04:26:40 \u29d6 2.000s\n'
+                u'    \u2514\u2500\u2500 app:action/2 \u21d2 succeeded '
+                u'2015-03-03 04:26:42\n'
                 u'\n'))
 
     def test_multiline_field(self):
@@ -870,10 +928,10 @@ class RenderTasksTests(TestCase):
             fd.getvalue(),
             ExactlyEquals(
                 u'f3a32bb3-ea6b-457c-aa99-08a3d0491ab4\n'
-                u'\u2514\u2500\u2500 app:action/1 \u21d2 started\n'
-                u'    \u251c\u2500\u2500 message: this is a\u23ce\n'
-                u'    \u2502   many line message\n'
-                u'    \u2514\u2500\u2500 timestamp: 1425356800\n\n'))
+                u'\u2514\u2500\u2500 app:action/1 \u21d2 started '
+                u'1425356800\n'
+                u'    \u2514\u2500\u2500 message: this is a\u23ce\n'
+                u'        many line message\n\n'))
 
     def test_multiline_field_limit(self):
         """
@@ -885,9 +943,9 @@ class RenderTasksTests(TestCase):
                               field_limit=1000),
             ExactlyEquals(
                 u'f3a32bb3-ea6b-457c-aa99-08a3d0491ab4\n'
-                u'\u2514\u2500\u2500 app:action/1 \u21d2 started\n'
-                u'    \u251c\u2500\u2500 message: this is a\u2026\n'
-                u'    \u2514\u2500\u2500 timestamp: 1425356800\n\n'))
+                u'\u2514\u2500\u2500 app:action/1 \u21d2 started '
+                u'1425356800\n'
+                u'    \u2514\u2500\u2500 message: this is a\u2026\n\n'))
 
     def test_field_limit(self):
         """
@@ -898,10 +956,10 @@ class RenderTasksTests(TestCase):
                               field_limit=5),
             ExactlyEquals(
                 u'cdeb220d-7605-4d5f-8341-1a170222e308\n'
-                u'\u2514\u2500\u2500 twisted:log/1\n'
+                u'\u2514\u2500\u2500 twisted:log/1 '
+                u'14253\u2026\n'
                 u'    \u251c\u2500\u2500 error: False\n'
-                u'    \u251c\u2500\u2500 message: Main \u2026\n'
-                u'    \u2514\u2500\u2500 timestamp: 14253\u2026\n\n'))
+                u'    \u2514\u2500\u2500 message: Main \u2026\n\n'))
 
     def test_ignored_keys(self):
         """
@@ -912,9 +970,9 @@ class RenderTasksTests(TestCase):
                               ignored_fields={u'action_type'}),
             ExactlyEquals(
                 u'f3a32bb3-ea6b-457c-aa99-08a3d0491ab4\n'
-                u'\u2514\u2500\u2500 app:action/1 \u21d2 started\n'
-                u'    \u251c\u2500\u2500 action_status: started\n'
-                u'    \u2514\u2500\u2500 timestamp: 1425356800\n\n'))
+                u'\u2514\u2500\u2500 app:action/1 \u21d2 started '
+                u'1425356800\n'
+                u'    \u2514\u2500\u2500 action_status: started\n\n'))
 
     def test_task_data(self):
         """
@@ -924,10 +982,10 @@ class RenderTasksTests(TestCase):
             self.render_tasks([message_task]),
             ExactlyEquals(
                 u'cdeb220d-7605-4d5f-8341-1a170222e308\n'
-                u'\u2514\u2500\u2500 twisted:log/1\n'
+                u'\u2514\u2500\u2500 twisted:log/1 '
+                u'1425356700\n'
                 u'    \u251c\u2500\u2500 error: False\n'
-                u'    \u251c\u2500\u2500 message: Main loop terminated.\n'
-                u'    \u2514\u2500\u2500 timestamp: 1425356700\n\n'))
+                u'    \u2514\u2500\u2500 message: Main loop terminated.\n\n'))
 
     def test_dict_data(self):
         """
@@ -937,10 +995,10 @@ class RenderTasksTests(TestCase):
             self.render_tasks([dict_action_task]),
             ExactlyEquals(
                 u'f3a32bb3-ea6b-457c-aa99-08a3d0491ab4\n'
-                u'\u2514\u2500\u2500 app:action/1 \u21d2 started\n'
-                u'    \u251c\u2500\u2500 some_data: \n'
-                u'    \u2502   \u2514\u2500\u2500 a: 42\n'
-                u'    \u2514\u2500\u2500 timestamp: 1425356800\n\n'))
+                u'\u2514\u2500\u2500 app:action/1 \u21d2 started '
+                u'1425356800\n'
+                u'    \u2514\u2500\u2500 some_data: \n'
+                u'        \u2514\u2500\u2500 a: 42\n\n'))
 
     def test_list_data(self):
         """
@@ -950,11 +1008,11 @@ class RenderTasksTests(TestCase):
             self.render_tasks([list_action_task]),
             ExactlyEquals(
                 u'f3a32bb3-ea6b-457c-aa99-08a3d0491ab4\n'
-                u'\u2514\u2500\u2500 app:action/1 \u21d2 started\n'
-                u'    \u251c\u2500\u2500 some_data: \n'
-                u'    \u2502   \u251c\u2500\u2500 0: a\n'
-                u'    \u2502   \u2514\u2500\u2500 1: b\n'
-                u'    \u2514\u2500\u2500 timestamp: 1425356800\n\n'))
+                u'\u2514\u2500\u2500 app:action/1 \u21d2 started '
+                u'1425356800\n'
+                u'    \u2514\u2500\u2500 some_data: \n'
+                u'        \u251c\u2500\u2500 0: a\n'
+                u'        \u2514\u2500\u2500 1: b\n\n'))
 
     def test_nested(self):
         """
@@ -964,10 +1022,10 @@ class RenderTasksTests(TestCase):
             self.render_tasks([action_task, nested_action_task]),
             ExactlyEquals(
                 u'f3a32bb3-ea6b-457c-aa99-08a3d0491ab4\n'
-                u'\u2514\u2500\u2500 app:action/1 \u21d2 started\n'
-                u'    \u251c\u2500\u2500 timestamp: 1425356800\n'
-                u'    \u2514\u2500\u2500 app:action:nest/1/1 \u21d2 started\n'
-                u'        \u2514\u2500\u2500 timestamp: 1425356900\n\n'))
+                u'\u2514\u2500\u2500 app:action/1 \u21d2 started '
+                u'1425356800\n'
+                u'    \u2514\u2500\u2500 app:action:nest/1/1 \u21d2 started '
+                u'1425356900\n\n'))
 
     def test_janky_message(self):
         """
@@ -978,11 +1036,11 @@ class RenderTasksTests(TestCase):
             self.render_tasks([janky_message_task]),
             ExactlyEquals(
                 u'cdeb220d-7605-4d5f-\u241b(08341-1a170222e308\n'
-                u'\u2514\u2500\u2500 M\u241b(0/1\n'
+                u'\u2514\u2500\u2500 M\u241b(0/1 '
+                u'1425356700\n'
                 u'    \u251c\u2500\u2500 er\u241bror: False\n'
-                u'    \u251c\u2500\u2500 mes\u240asage: '
-                u'Main loop\u241b(0terminated.\n'
-                u'    \u2514\u2500\u2500 timestamp: 1425356700\n\n'))
+                u'    \u2514\u2500\u2500 mes\u240asage: '
+                u'Main loop\u241b(0terminated.\n\n'))
 
     def test_janky_action(self):
         """
@@ -993,11 +1051,12 @@ class RenderTasksTests(TestCase):
             self.render_tasks([janky_action_task]),
             ExactlyEquals(
                 u'f3a32bb3-ea6b-457c-\u241b(0aa99-08a3d0491ab4\n'
-                u'\u2514\u2500\u2500 A\u241b(0/1 \u21d2 started\n'
+                u'\u2514\u2500\u2500 A\u241b(0/1 \u21d2 started '
+                u'1425356800\u241b(0\n'
                 u'    \u251c\u2500\u2500 \u241b(0: \n'
                 u'    \u2502   \u2514\u2500\u2500 \u241b(0: nope\n'
-                u'    \u251c\u2500\u2500 mes\u240asage: hello\u241b(0world\n'
-                u'    \u2514\u2500\u2500 timestamp: 1425356800\u241b(0\n\n'))
+                u'    \u2514\u2500\u2500 mes\u240asage: hello\u241b(0world\n\n'
+            ))
 
     def test_colorize(self):
         """
@@ -1009,15 +1068,15 @@ class RenderTasksTests(TestCase):
             ExactlyEquals(
                 u'\n'.join([
                     colors.root(u'f3a32bb3-ea6b-457c-aa99-08a3d0491ab4'),
-                    u'\u2514\u2500\u2500 {}/1 \u21d2 {}'.format(
+                    u'\u2514\u2500\u2500 {}/1 \u21d2 {} {} {} {}'.format(
                         colors.parent(u'app:action'),
-                        colors.success(u'started')),
-                    u'    \u251c\u2500\u2500 {}: {}'.format(
-                        colors.prop(u'timestamp'), u'1425356800'),
-                    u'    \u2514\u2500\u2500 {}/2 \u21d2 {}'.format(
+                        colors.success(u'started'),
+                        colors.timestamp(u'1425356800'),
+                        HOURGLASS,
+                        colors.duration(u'2')),
+                    u'    \u2514\u2500\u2500 {}/2 \u21d2 {} {}'.format(
                         colors.parent(u'app:action'),
-                        colors.success(u'succeeded')),
-                    u'        \u2514\u2500\u2500 {}: {}'.format(
-                        colors.prop('timestamp'), u'1425356800'),
+                        colors.success(u'succeeded'),
+                        colors.timestamp(u'1425356802')),
                     u'\n',
                 ])))